Convert synthesized Android motion events to long and not to int. (flutter/engine#6103)

+++ engine/src/flutter/shell/platform/android/io/flutter/plugin/platform/PlatformViewsController.java
@@ -254,8 +254,8 @@ public class PlatformViewsController implements MethodChannel.MethodCallHandler 
         }
 
         MotionEvent event = MotionEvent.obtain(
-                downTime.intValue(),
-                eventTime.intValue(),
+                downTime.longValue(),
+                eventTime.longValue(),
                 action,
                 pointerCount,
                 pointerProperties,
